Add api/ prefix to API URLs
This is a fix for solving pagination links in the API.

Note that this is breaking upstream merging capabilities with the Helsinki setup as they do not have such prefix in their codebase. The reason for this is that they are using uwsgi+kong to handle such URL prefixing, something that would currently take too long for us to solve.
In the future, if this code is to be merge, solving this issue in another way would be recommended.